+++ asapdiscovery-docking/asapdiscovery/docking/workflows/large_scale_docking.py
@@ -120,9 +120,10 @@ def large_scale_docking_workflow(inputs: LargeScaleDockingInputs):
     """
 
     output_dir = inputs.output_dir
-    if output_dir.exists():
+    if output_dir.exists() and inputs.overwrite:
         rmtree(output_dir)
-    output_dir.mkdir()
+    else:
+        output_dir.mkdir(exist_ok=True, parents=True)
 
     logger = FileLogger(
         inputs.logname,  # default root logger so that dask logging is forwarded
